+++ internal/driver/sqlite/dialect.go
@@ -83,7 +83,7 @@ func (d *Dialect) ColumnListForSelect(cols, colTypes []string, targetDBType stri
 func (d *Dialect) BuildKeysetQuery(cols, pkCol, schema, table, _ string, hasMaxPK bool, dateFilter *driver.DateFilter) string {
 	dateClause := ""
 	if dateFilter != nil {
-		dateClause = fmt.Sprintf(" AND %s >= ?", d.QuoteIdentifier(dateFilter.Column))
+		dateClause = fmt.Sprintf(" AND %s > ?", d.QuoteIdentifier(dateFilter.Column))
 	}
 
 	qualifiedTable := d.QualifyTable(schema, table)
@@ -125,7 +125,7 @@ func (d *Dialect) BuildRowNumberQuery(cols, orderBy, schema, table, _ string, da
 
 	whereClause := ""
 	if dateFilter != nil {
-		whereClause = fmt.Sprintf(" WHERE %s >= ?", d.QuoteIdentifier(dateFilter.Column))
+		whereClause = fmt.Sprintf(" WHERE %s > ?", d.QuoteIdentifier(dateFilter.Column))
 	}
 
 	return fmt.Sprintf(`